Improved sync bundle template generation.
Added a new config item specific to branch policy named
`syncexcludebundles` that will exclude the files from this
bundle in the paths in sync automation.

Also changed config and test - it still doesn't generate the
exact sync automation - but generates a  file that's
functionally same as the currently working one.