Added goreleaser el7 templates.
In tyk-ci, the el7 templates were managed by a special case for el7 file
in the generation script itself, which avoided having to maintain  a lot of
redundant templates. But here we've offloaded the el7 changes to templates
mainly due to the inability to include templates by calling filenames.
This necessitates keeping some redundant templates, but avoids having to
maintain a special case in the code for `el7` files/templates.