Refactor: Adds repo policy rendering for git pr bundles.
This adds a `print-policy` sub command to git that
can print the release branches policy for the given repo.
Contains some refactoring that was needed to achieve this - the
gitRepo object that renders the pr bundles had no way of knowing
policy info - this adds a RepoPolicy key to the git repo struct.

As policy package is completely isolated from git - this removes
some of the git references from policy and avoids an import cycle.