Fix git push
gromit git push was failing if the local branch to be pushed
contained "/" in the branch name. This was due to a bug we
had in our Branch() function.
This commit fixes it by updating the Branch() function to
avoid using string manipulation to get the short ref name
for branch.